+++ src/sparsezoo/analyze_v2/model_validator/sparsity_analysis_model.py
@@ -14,7 +14,7 @@
 
 from pydantic import BaseModel, Field, validator
 
-from sparsezoo.analyze_v2.model_validators.utils import type_validator
+from sparsezoo.analyze_v2.model_validator.utils import type_validator
 
 
 class SparsityAnalysisModel(BaseModel):
